Refactor and extract bep5 and bep6 shared code; Remove unnecessary differences between versions
