+++ BepInEx.KeyboardShortcut/AssemblyInfo.cs
