+++ ConfigurationManager.Shared/ConfigurationManager.Shared.projitems
@@ -1,0 +1,23 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>28e965b5-225f-4ad4-938f-2d6f8c20c73f</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>ConfigurationManager.Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)AssemblyInfo.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ConfigSettingEntry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ConfigurationManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)PropertySettingEntry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SettingEntryBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SettingFieldDrawer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\ComboBox.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\ImguiUtils.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\Utilities.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ValueChangedEventArgs.cs" />
+  </ItemGroup>
+</Project>
