+++ ConfigurationManager.Shared/PropertySettingEntry.cs
@@ -2,6 +2,10 @@
 using System.Reflection;
 using BepInEx;
 
+#if IL2CPP
+using BaseUnityPlugin = BepInEx.PluginInfo;
+#endif
+
 namespace ConfigurationManager
 {
     internal class PropertySettingEntry : SettingEntryBase
