+++ ConfigurationManager.Shared/ValueChangedEventArgs.cs
@@ -2,7 +2,6 @@
 // Copyright 2018 GNU General Public License v3.0
 
 using System;
-
 namespace ConfigurationManager
 {
     /// <summary>
