--- ConfigurationManager/Properties/AssemblyInfo.cs
@@ -1,15 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyProduct("BepInEx.ConfigurationManager")]
-[assembly: AssemblyTitle("BepInEx.ConfigurationManager")]
-
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("https://github.com/BepInEx/BepInEx.ConfigurationManager")]
-[assembly: AssemblyCopyright("Copyright © 2019")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-
-[assembly: AssemblyVersion(ConfigurationManager.ConfigurationManager.Version)]
